requirements.txt
sqlalchemy>=2.0.0
faker>=19.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0 

beautifulsoup4>=4.12.0